@@ -1,30 +1,30 @@
 name: geocoder_buddy
 description: Forward, Reverse geocoding and Search Address without any apikey.
-version: 1.0.1
+version: 1.0.2
 homepage: https://github.com/flutterbuddy1/geocoder_buddy
 repository: https://github.com/flutterbuddy1/geocoder_buddy
 issue_tracker: https://github.com/flutterbuddy1/geocoder_buddy/issues
 
 environment:
-  sdk: ">=2.17.5 <3.0.0"
+  sdk: ">=3.4.0 <4.0.0"
   flutter: ">=1.17.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  http: ^0.13.4
-  dartdoc: ^6.0.0
+  http: ^1.2.2
+  dartdoc: ^8.1.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_lints: ^2.0.0
+  flutter_lints: ^5.0.0
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
 
 # The following section is specific to Flutter packages.
-flutter:
+# flutter:
 
   # To add assets to your package, add an assets section, like this:
   # assets:
